Move court case related helper methods to a separate module
Once we synced agreements and court_cases from UH, we can swap these easily with a new module that uses the court cases
in MA
