+++ lib/hackney/income/tenancy_classification/v2/helpers.rb
@@ -3,9 +3,12 @@ module Hackney
     module TenancyClassification
       module V2
         module Helpers
-          # Replace this with Hackney::Income::TenancyClassification::V2::MAAgreementHelpers
+          # Replace these with:
+          # include Hackney::Income::TenancyClassification::V2::Helpers::MAAgreementHelpers
+          # include Hackney::Income::TenancyClassification::V2::Helpers::MACourtCaseHelpers
           # when agreements synced from UH
           include Hackney::Income::TenancyClassification::V2::Helpers::UHAgreementHelpers
+          include Hackney::Income::TenancyClassification::V2::Helpers::UHCourtCaseHelpers
 
           def case_paused?
             @case_priority.paused?
@@ -15,40 +18,8 @@ module Hackney
             @criteria.eviction_date.present?
           end
 
-          def court_warrant_active?
-            return false if @criteria.court_outcome.blank?
-
-            if @criteria.court_outcome.in?([
-              Hackney::Tenancy::CourtOutcomeCodes::ADJOURNED_GENERALLY,
-              Hackney::Tenancy::CourtOutcomeCodes::ADJOURNED_ON_TERMS,
-              Hackney::Tenancy::CourtOutcomeCodes::ADJOURNED_ON_TERMS_SECONDARY,
-              Hackney::Tenancy::CourtOutcomeCodes::SUSPENDED_POSSESSION,
-              Hackney::Tenancy::CourtOutcomeCodes::OUTRIGHT_POSSESSION_FORTHWITH,
-              Hackney::Tenancy::CourtOutcomeCodes::OUTRIGHT_POSSESSION_WITH_DATE
-            ])
-              return @criteria.courtdate.blank? || @criteria.courtdate > 6.years.ago
-            end
-
-            false
-          end
-
-          def court_date_in_future?
-            @criteria.courtdate.present? && @criteria.courtdate.future?
-          end
-
           def should_prevent_action?
             case_has_eviction_date? || court_date_in_future? || case_paused?
-          end
-
-          def no_court_date?
-            @criteria.courtdate.blank?
-          end
-
-          def court_outcome_missing?
-            return false if court_date_in_future?
-            return false if no_court_date?
-
-            @criteria.court_outcome.blank?
           end
 
           def last_communication_older_than?(date)
